move CompactBytes to the 'growable' moddule
